pyproject.toml
[tool.poetry]
name = "codeshepherds"
version = "0.8.0"
description = "Webpage of codeshepherds"
authors = ["Leopoldo Parra <lparra.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.*"
celery = "*"
channels = "*"
channels-redis = "*"
colorama = "*"
coreapi = "*"
cryptography = "*"
django = "*"
django-bootstrap4 = "*"
django-cors-headers = "*"
django-crispy-forms = "*"
django-icons = "*"
django-ipware = "*"
django-redis-cache = "*"
django-widget-tweaks = "*"
djangorestframework = "*"
dropbox = "*"
geoip2 = "*"
gunicorn = "*"
hiredis = "*"
pillow = "*"
psycopg2-binary = "*"
pygraphviz = "*"
redis = "*"
sentry-sdk = "*"
uvicorn = "*"
whitenoise = "*"
cs-django-thumbnails = "*"
Twisted = {extras = ["tls", "http2"], version = "*"}
django-cleanup = "*"
django-simple-history = "*"
django-admin-honeypot = "*"
sendgrid = "*"

[tool.poetry.dev-dependencies]
coverage = "*"
flake8 = "*"
django-debug-toolbar = "*"
django-extensions = "*"
werkzeug = "*"
selenium = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
